add test to make sure querying endpoints from a service is not allowed
As this type of query is not allowed we make sure it actually is with a
new test.

Quoting the MCS-API KEP with the explanation of why this is not allowed:
While we reserve the form <clusterid>.<svc>.<ns>.svc.clusterset.local. for
possible future use, both ClusterSetIP Services and Multicluster Headless Services
are specified to explicitly disallow using this form to create DNS records that
target all 1+N backends in a specific cluster.

Signed-off-by: Arthur Outhenin-Chalandre <arthur@cri.epita.fr>