object: fix object cleanup (#27)
This commit add object cleanup to be consistent with similar code in
the kubernetes plugin.

It also does a clone of the EndpointSlice labels so that it's not erased by
the function `EndpointSliceToEndpoints` from the kubernetes plugin fixing an
important bug in the multicluster code as it was previously not able to
retrieve any labels which essentially prevented any interaction with headless
services.

Signed-off-by: Arthur Outhenin-Chalandre <arthur@cri.epita.fr>